+++ cookbooks/columbus_java/attributes/default.rb
@@ -1,2 +1,0 @@
-default['java']['jdk_version'] = '7'
-