+++ cookbooks/java/recipes/openjdk.rb
@@ -45,7 +45,7 @@ end
 if node['platform'] == 'ubuntu'
   include_recipe 'apt'
   apt_repository 'openjdk-r-ppa' do
-    uri 'ppa:openjdk-r/ppa'
+    uri 'ppa:openjdk-r'
     distribution node['lsb']['codename']
   end
 end