+++ cookbooks/test1/recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: jdk
+# Cookbook Name:: test1
 # Recipe:: default
 #
 # Copyright 2016, YOUR_COMPANY_NAME