+++ cookbooks/mindtree/CHANGELOG.md
@@ -1,9 +1,9 @@
-# jdk CHANGELOG
+# mindtree CHANGELOG
 
-This file is used to list changes made in each version of the jdk cookbook.
+This file is used to list changes made in each version of the mindtree cookbook.
 
 ## 0.1.0
-- [your_name] - Initial release of jdk
+- [your_name] - Initial release of mindtree
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.